refactor: update imports in react files to use dash-case
