+++ src/react/index.ts
@@ -1,3 +1,3 @@
-export { FireproofCtx, useFireproof } from "./useFireproof.js";
+export { FireproofCtx, useFireproof } from "./use-fireproof.js";
 export * from "./types.js";
 export * from "./img-file.js";
